+++ presets/special-numbers.js
@@ -1,11 +1,13 @@
 import nan from '../types/nan.js';
 import infinity from '../types/infinity.js';
-import NegativeInfinity from '../types/negative-infinity.js';
+import negativeInfinity from '../types/negative-infinity.js';
+import negativeZero from '../types/negative-zero.js';
 
 const specialNumbers = [
     nan,
     infinity,
-    NegativeInfinity
+    negativeInfinity,
+    negativeZero
 ];
 
 export default specialNumbers;